Add publish/unpublish functionality to admin panel.
Now all pushed content packs are put into quarantine first. An admin has to publish them
from the panel. The admin has also the option to inspect each quarantined content pack
and open a simple preview.

As of now, this is the default behavior. A configuration attribute will be added, which
will determine whether or not "quarantine-mode" should be used.